+++ sources/org/telegram/tgnet/TLRPC$TL_decryptedMessageMediaAudio_layer8.java
@@ -5,7 +5,7 @@ public class TLRPC$TL_decryptedMessageMediaAudio_layer8 extends TLRPC$TL_decrypt
 
     public void readParams(AbstractSerializedData abstractSerializedData, boolean z) {
         this.duration = abstractSerializedData.readInt32(z);
-        this.size = abstractSerializedData.readInt32(z);
+        this.size = (long) abstractSerializedData.readInt32(z);
         this.key = abstractSerializedData.readByteArray(z);
         this.iv = abstractSerializedData.readByteArray(z);
     }
@@ -13,7 +13,7 @@ public class TLRPC$TL_decryptedMessageMediaAudio_layer8 extends TLRPC$TL_decrypt
     public void serializeToStream(AbstractSerializedData abstractSerializedData) {
         abstractSerializedData.writeInt32(constructor);
         abstractSerializedData.writeInt32(this.duration);
-        abstractSerializedData.writeInt32(this.size);
+        abstractSerializedData.writeInt32((int) this.size);
         abstractSerializedData.writeByteArray(this.key);
         abstractSerializedData.writeByteArray(this.iv);
     }
