+++ sources/org/telegram/tgnet/TLRPC$TL_decryptedMessageMediaVideo_layer17.java
@@ -12,7 +12,7 @@ public class TLRPC$TL_decryptedMessageMediaVideo_layer17 extends TLRPC$TL_decryp
         this.mime_type = abstractSerializedData.readString(z);
         this.w = abstractSerializedData.readInt32(z);
         this.h = abstractSerializedData.readInt32(z);
-        this.size = abstractSerializedData.readInt32(z);
+        this.size = (long) abstractSerializedData.readInt32(z);
         this.key = abstractSerializedData.readByteArray(z);
         this.iv = abstractSerializedData.readByteArray(z);
     }
@@ -26,7 +26,7 @@ public class TLRPC$TL_decryptedMessageMediaVideo_layer17 extends TLRPC$TL_decryp
         abstractSerializedData.writeString(this.mime_type);
         abstractSerializedData.writeInt32(this.w);
         abstractSerializedData.writeInt32(this.h);
-        abstractSerializedData.writeInt32(this.size);
+        abstractSerializedData.writeInt32((int) this.size);
         abstractSerializedData.writeByteArray(this.key);
         abstractSerializedData.writeByteArray(this.iv);
     }
