+++ sources/org/telegram/tgnet/TLRPC$TL_documentEncrypted.java
@@ -8,7 +8,7 @@ public class TLRPC$TL_documentEncrypted extends TLRPC$Document {
         this.access_hash = abstractSerializedData.readInt64(z);
         this.date = abstractSerializedData.readInt32(z);
         this.mime_type = abstractSerializedData.readString(z);
-        this.size = abstractSerializedData.readInt32(z);
+        this.size = (long) abstractSerializedData.readInt32(z);
         this.thumbs.add(TLRPC$PhotoSize.TLdeserialize(0, 0, 0, abstractSerializedData, abstractSerializedData.readInt32(z), z));
         this.dc_id = abstractSerializedData.readInt32(z);
         int readInt32 = abstractSerializedData.readInt32(z);
@@ -37,7 +37,7 @@ public class TLRPC$TL_documentEncrypted extends TLRPC$Document {
         abstractSerializedData.writeInt64(this.access_hash);
         abstractSerializedData.writeInt32(this.date);
         abstractSerializedData.writeString(this.mime_type);
-        abstractSerializedData.writeInt32(this.size);
+        abstractSerializedData.writeInt32((int) this.size);
         this.thumbs.get(0).serializeToStream(abstractSerializedData);
         abstractSerializedData.writeInt32(this.dc_id);
         abstractSerializedData.writeInt32(NUM);
