+++ sources/org/telegram/tgnet/TLRPC$TL_upload_getCdnFile.java
@@ -4,7 +4,7 @@ public class TLRPC$TL_upload_getCdnFile extends TLObject {
     public static int constructor = NUM;
     public byte[] file_token;
     public int limit;
-    public int offset;
+    public long offset;
 
     public TLObject deserializeResponse(AbstractSerializedData abstractSerializedData, int i, boolean z) {
         return TLRPC$upload_CdnFile.TLdeserialize(abstractSerializedData, i, z);
@@ -13,7 +13,7 @@ public class TLRPC$TL_upload_getCdnFile extends TLObject {
     public void serializeToStream(AbstractSerializedData abstractSerializedData) {
         abstractSerializedData.writeInt32(constructor);
         abstractSerializedData.writeByteArray(this.file_token);
-        abstractSerializedData.writeInt32(this.offset);
+        abstractSerializedData.writeInt64(this.offset);
         abstractSerializedData.writeInt32(this.limit);
     }
 }
